fix interface, logout function
