+++ src/layout/Header.tsx
@@ -1,31 +1,25 @@
 import { useReactQueryDelete } from '@/api/http';
 import { LOGIN_PAGE_URL } from '@/constants/utl';
+import { initUserData } from '@/store/userSlice';
 import { centerRowStyles, spaceBetweenRowStyles } from '@/styles/flexModules';
+import { IHeaderReduxState } from '@/types/interfaces/commons';
 import styled from '@emotion/styled';
 import Head from 'next/head';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import { useSelector } from 'react-redux';
-
-interface IHeaderReduxState {
-    user: {
-        accessToken: string;
-        adminInfo: {
-            adminName: string;
-        };
-        isLogin: boolean;
-    };
-}
+import { useDispatch, useSelector } from 'react-redux';
 
 export default function Header() {
     const { adminInfo, isLogin } = useSelector(
         (state: IHeaderReduxState) => state.user,
     );
+    const dispatch = useDispatch();
 
     const { mutation: logoutMutation, isLoading } = useReactQueryDelete({
         url: '/admin/logout',
         onSuccess: () => {
+            dispatch(initUserData());
             router.push(LOGIN_PAGE_URL);
         },
     });
