+++ src/types/interfaces/commons.ts
@@ -136,3 +136,23 @@ export interface ICalenderNavProps {
     isMonthFirstDay: string;
     setIsMonthFirstDay: (changeMonth: string) => void;
 }
+
+export interface IUserInitialState {
+    accessToken?: string;
+    refreshToken?: string;
+    isLogin?: boolean;
+    adminInfo?: {
+        adminId: string;
+        adminName: string;
+    };
+}
+
+export interface IHeaderReduxState {
+    user: {
+        accessToken: string;
+        adminInfo: {
+            adminName: string;
+        };
+        isLogin: boolean;
+    };
+}
